Moved clock_reduction to own module, we have begun moving clocks to TransitionSystem. Cleaned up in extract_system_rep.
