+++ src/transition_systems/mod.rs
@@ -1,5 +1,6 @@
 #[macro_use]
 pub(crate) mod common;
+pub mod clock_reduction;
 mod compiled_component;
 pub mod compiled_update;
 mod composition;
